+++ CHANGELOG.md
@@ -2,6 +2,9 @@
 
 ## Unreleased
 
+- Add MemoryRegionInfo bindings.
+- Add APISupport module for feature detection.
+
 ## 0.1.0 - 2025-12-30
 
 - Initial release.